[ALL] Refactor, organization, added search to CLI
