+++ owmods_core/Cargo.toml
@@ -28,6 +28,7 @@ uuid = { version = "1.3.0", features = ["v4", "fast-rng"] }
 typeshare = "1.0.0"
 lazy_static = "1.4.0"
 tempfile = "3.5.0"
+rust-fuzzy-search = "0.1.1"
 
 [dev-dependencies]
 tokio-test = "0.4.2"
