+++ owmods_core/src/download.rs
@@ -17,7 +17,8 @@ use crate::{
     config::Config,
     db::{LocalDatabase, RemoteDatabase},
     file::{check_file_matches_paths, create_all_parents, fix_json},
-    mods::{get_paths_to_preserve, LocalMod, ModManifest, RemoteMod},
+    mods::local::{get_paths_to_preserve, LocalMod, ModManifest},
+    mods::remote::RemoteMod,
     progress::{ProgressAction, ProgressBar, ProgressType},
     remove::remove_old_mod_files,
     toggle::generate_config,
@@ -433,10 +434,13 @@ pub async fn install_mod_from_db(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::file::serialize_to_json;
-    use crate::mods::UnsafeLocalMod;
-    use crate::test_utils::{get_test_file, make_test_dir};
+    use crate::{
+        file::serialize_to_json,
+        mods::local::UnsafeLocalMod,
+        test_utils::{get_test_file, make_test_dir},
+    };
     use std::fs::read_to_string;
+
     const TEST_URL: &str =
         "https://github.com/Bwc9876/OW-TimeSaver/releases/download/1.1.1/Bwc9876.TimeSaver.zip";
