+++ owmods_core/src/mods/mod.rs
@@ -1,0 +1,5 @@
+/// Definitions for local (installed) mods
+pub mod local;
+
+/// Definitions for remote (in the database) mods
+pub mod remote;
