+++ owmods_gui/backend/Cargo.toml
@@ -15,7 +15,6 @@ tauri = { version = "1.3.0", features = ["app-all", "dialog-ask", "dialog-confir
 anyhow = "1.0.69"
 tokio = { version = "1.25.0", features = ["sync", "macros"] }
 log = { version = "0.4.17", features = ["std", "serde"] }
-rust-fuzzy-search = "0.1.1"
 typeshare = "1.0.0"
 regex = "1.7.1"
 time = { version = "0.3.15", features = ["macros"] }
